+++ pkg/BUILD.bazel
@@ -1163,6 +1163,7 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/idalloc:idalloc_test",
     "//pkg/kv/kvserver/intentresolver:intentresolver",
     "//pkg/kv/kvserver/intentresolver:intentresolver_test",
+    "//pkg/kv/kvserver/kvadmission:kvadmission",
     "//pkg/kv/kvserver/kvserverbase:kvserverbase",
     "//pkg/kv/kvserver/kvserverpb:kvserverpb",
     "//pkg/kv/kvserver/liveness/livenesspb:livenesspb",
@@ -2480,6 +2481,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/kv/kvserver/gc:get_x_data",
     "//pkg/kv/kvserver/idalloc:get_x_data",
     "//pkg/kv/kvserver/intentresolver:get_x_data",
+    "//pkg/kv/kvserver/kvadmission:get_x_data",
     "//pkg/kv/kvserver/kvserverbase:get_x_data",
     "//pkg/kv/kvserver/kvserverpb:get_x_data",
     "//pkg/kv/kvserver/liveness:get_x_data",
