+++ pkg/kv/kvserver/replica_application_state_machine.go
@@ -17,12 +17,12 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/apply"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/ctpb"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvadmission"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverbase"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverpb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage"
 	"github.com/cockroachdb/cockroach/pkg/storage/enginepb"
-	"github.com/cockroachdb/cockroach/pkg/util/admission"
 	"github.com/cockroachdb/cockroach/pkg/util/envutil"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
@@ -56,20 +56,7 @@ type applyCommittedEntriesStats struct {
 	stateAssertions         int
 	numEmptyEntries         int
 	numConfChangeEntries    int
-	followerStoreWriteBytes followerStoreWriteBytes
-}
-
-// followerStoreWriteBytes captures stats about writes done to a store by a
-// replica that is not the leaseholder. These are used for admission control.
-type followerStoreWriteBytes struct {
-	numEntries int64
-	admission.StoreWorkDoneInfo
-}
-
-func (f *followerStoreWriteBytes) merge(from followerStoreWriteBytes) {
-	f.numEntries += from.numEntries
-	f.WriteBytes += from.WriteBytes
-	f.IngestedBytes += from.IngestedBytes
+	followerStoreWriteBytes kvadmission.FollowerStoreWriteBytes
 }
 
 // nonDeterministicFailure is an error type that indicates that a state machine
@@ -449,7 +436,7 @@ type replicaAppBatch struct {
 	emptyEntries            int
 	mutations               int
 	start                   time.Time
-	followerStoreWriteBytes followerStoreWriteBytes
+	followerStoreWriteBytes kvadmission.FollowerStoreWriteBytes
 
 	// Reused by addAppliedStateKeyToBatch to avoid heap allocations.
 	asAlloc enginepb.RangeAppliedState
@@ -550,7 +537,7 @@ func (b *replicaAppBatch) Stage(
 	// nils the AddSSTable field.
 	if !cmd.IsLocal() {
 		writeBytes, ingestedBytes := cmd.getStoreWriteByteSizes()
-		b.followerStoreWriteBytes.numEntries++
+		b.followerStoreWriteBytes.NumEntries++
 		b.followerStoreWriteBytes.WriteBytes += writeBytes
 		b.followerStoreWriteBytes.IngestedBytes += ingestedBytes
 	}
@@ -1063,7 +1050,7 @@ func (b *replicaAppBatch) recordStatsOnCommit() {
 	b.sm.stats.entriesProcessedBytes += b.entryBytes
 	b.sm.stats.numEmptyEntries += b.emptyEntries
 	b.sm.stats.batchesProcessed++
-	b.sm.stats.followerStoreWriteBytes.merge(b.followerStoreWriteBytes)
+	b.sm.stats.followerStoreWriteBytes.Merge(b.followerStoreWriteBytes)
 
 	elapsed := timeutil.Since(b.start)
 	b.r.store.metrics.RaftCommandCommitLatency.RecordValue(elapsed.Nanoseconds())
