+++ pkg/kv/kvserver/store.go
@@ -40,6 +40,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/sidetransport"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/idalloc"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/intentresolver"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvadmission"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvserverpb"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/multiqueue"
@@ -61,7 +62,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/storage/enginepb"
 	"github.com/cockroachdb/cockroach/pkg/util/admission"
 	"github.com/cockroachdb/cockroach/pkg/util/admission/admissionpb"
-	"github.com/cockroachdb/cockroach/pkg/util/buildutil"
 	"github.com/cockroachdb/cockroach/pkg/util/contextutil"
 	"github.com/cockroachdb/cockroach/pkg/util/ctxgroup"
 	"github.com/cockroachdb/cockroach/pkg/util/envutil"
@@ -84,7 +84,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/logtags"
-	"github.com/cockroachdb/pebble"
 	"github.com/cockroachdb/redact"
 	prometheusgo "github.com/prometheus/client_model/go"
 	"go.etcd.io/etcd/raft/v3"
@@ -1105,7 +1104,7 @@ type StoreConfig struct {
 	SpanConfigSubscriber spanconfig.KVSubscriber
 
 	// KVAdmissionController is an optional field used for admission control.
-	KVAdmissionController KVAdmissionController
+	KVAdmissionController kvadmission.Controller
 
 	// SchedulerLatencyListener listens in on scheduling latencies, information
 	// that's then used to adjust various admission control components (like how
@@ -3887,329 +3886,3 @@ func min(a, b int) int {
 	}
 	return b
 }
-
-// elasticCPUDurationPerExportRequest controls how many CPU tokens are allotted
-// for each export request.
-var elasticCPUDurationPerExportRequest = settings.RegisterDurationSetting(
-	settings.SystemOnly,
-	"kvadmission.elastic_cpu.duration_per_export_request",
-	"controls how many CPU tokens are allotted for each export request",
-	admission.MaxElasticCPUDuration,
-	func(duration time.Duration) error {
-		if duration < admission.MinElasticCPUDuration {
-			return fmt.Errorf("minimum CPU duration allowed per export request is %s, got %s",
-				admission.MinElasticCPUDuration, duration)
-		}
-		if duration > admission.MaxElasticCPUDuration {
-			return fmt.Errorf("maximum CPU duration allowed per export request is %s, got %s",
-				admission.MaxElasticCPUDuration, duration)
-		}
-		return nil
-	},
-)
-
-// KVAdmissionController provides admission control for the KV layer.
-type KVAdmissionController interface {
-	// AdmitKVWork must be called before performing KV work.
-	// BatchRequest.AdmissionHeader and BatchRequest.Replica.StoreID must be
-	// populated for admission to work correctly. If err is non-nil, the
-	// returned handle can be ignored. If err is nil, AdmittedKVWorkDone must be
-	// called after the KV work is done executing.
-	AdmitKVWork(
-		ctx context.Context, tenantID roachpb.TenantID, ba *roachpb.BatchRequest,
-	) (AdmissionHandle, error)
-	// AdmittedKVWorkDone is called after the admitted KV work is done
-	// executing.
-	AdmittedKVWorkDone(AdmissionHandle, *StoreWriteBytes)
-	// SetTenantWeightProvider is used to set the provider that will be
-	// periodically polled for weights. The stopper should be used to terminate
-	// the periodic polling.
-	SetTenantWeightProvider(provider TenantWeightProvider, stopper *stop.Stopper)
-	// SnapshotIngested informs admission control about a range snapshot
-	// ingestion.
-	SnapshotIngested(storeID roachpb.StoreID, ingestStats pebble.IngestOperationStats)
-	// FollowerStoreWriteBytes informs admission control about writes
-	// replicated to a raft follower, that have not been subject to admission
-	// control.
-	FollowerStoreWriteBytes(storeID roachpb.StoreID, followerWriteBytes followerStoreWriteBytes)
-}
-
-// TenantWeightProvider can be periodically asked to provide the tenant
-// weights.
-type TenantWeightProvider interface {
-	GetTenantWeights() TenantWeights
-}
-
-// TenantWeights contains the various tenant weights.
-type TenantWeights struct {
-	// Node is the node level tenant ID => weight.
-	Node map[uint64]uint32
-	// Stores contains the per-store tenant weights.
-	Stores []TenantWeightsForStore
-}
-
-// TenantWeightsForStore contains the tenant weights for a store.
-type TenantWeightsForStore struct {
-	roachpb.StoreID
-	// Weights is tenant ID => weight.
-	Weights map[uint64]uint32
-}
-
-// KVAdmissionControllerImpl implements KVAdmissionController interface.
-type KVAdmissionControllerImpl struct {
-	// Admission control queues and coordinators. All three should be nil or
-	// non-nil.
-	kvAdmissionQ        *admission.WorkQueue
-	storeGrantCoords    *admission.StoreGrantCoordinators
-	elasticCPUWorkQueue *admission.ElasticCPUWorkQueue
-	settings            *cluster.Settings
-	every               log.EveryN
-}
-
-var _ KVAdmissionController = &KVAdmissionControllerImpl{}
-
-// AdmissionHandle groups data around some piece admitted work. Depending on the
-// type of work, it holds (a) references to specific work queues, (b) state
-// needed to inform said work queues of what work was done after the fact, and
-// (c) information around how much work a request is allowed to do (used for
-// cooperative scheduling with elastic CPU granters).
-//
-// TODO(irfansharif): Consider moving KVAdmissionController and adjacent types
-// into a kvserver/kvadmission package.
-type AdmissionHandle struct {
-	tenantID             roachpb.TenantID
-	storeAdmissionQ      *admission.StoreWorkQueue
-	storeWorkHandle      admission.StoreWorkHandle
-	ElasticCPUWorkHandle *admission.ElasticCPUWorkHandle
-
-	callAdmittedWorkDoneOnKVAdmissionQ bool
-}
-
-// MakeKVAdmissionController returns a KVAdmissionController. All three
-// parameters must together be nil or non-nil.
-func MakeKVAdmissionController(
-	kvAdmissionQ *admission.WorkQueue,
-	elasticCPUWorkQueue *admission.ElasticCPUWorkQueue,
-	storeGrantCoords *admission.StoreGrantCoordinators,
-	settings *cluster.Settings,
-) KVAdmissionController {
-	return &KVAdmissionControllerImpl{
-		kvAdmissionQ:        kvAdmissionQ,
-		storeGrantCoords:    storeGrantCoords,
-		elasticCPUWorkQueue: elasticCPUWorkQueue,
-		settings:            settings,
-		every:               log.Every(10 * time.Second),
-	}
-}
-
-// AdmitKVWork implements the KVAdmissionController interface.
-//
-// TODO(irfansharif): There's a fair bit happening here and there's no test
-// coverage. Fix that.
-func (n *KVAdmissionControllerImpl) AdmitKVWork(
-	ctx context.Context, tenantID roachpb.TenantID, ba *roachpb.BatchRequest,
-) (handle AdmissionHandle, retErr error) {
-	ah := AdmissionHandle{tenantID: tenantID}
-	if n.kvAdmissionQ == nil {
-		return ah, nil
-	}
-
-	bypassAdmission := ba.IsAdmin()
-	source := ba.AdmissionHeader.Source
-	if !roachpb.IsSystemTenantID(tenantID.ToUint64()) {
-		// Request is from a SQL node.
-		bypassAdmission = false
-		source = roachpb.AdmissionHeader_FROM_SQL
-	}
-	if source == roachpb.AdmissionHeader_OTHER {
-		bypassAdmission = true
-	}
-	// TODO(abaptist): Revisit and deprecate this setting in v23.1.
-	if admission.KVBulkOnlyAdmissionControlEnabled.Get(&n.settings.SV) {
-		if admissionpb.WorkPriority(ba.AdmissionHeader.Priority) >= admissionpb.NormalPri {
-			bypassAdmission = true
-		}
-	}
-
-	createTime := ba.AdmissionHeader.CreateTime
-	if !bypassAdmission && createTime == 0 {
-		// TODO(sumeer): revisit this for multi-tenant. Specifically, the SQL use
-		// of zero CreateTime needs to be revisited. It should use high priority.
-		createTime = timeutil.Now().UnixNano()
-	}
-	admissionInfo := admission.WorkInfo{
-		TenantID:        tenantID,
-		Priority:        admissionpb.WorkPriority(ba.AdmissionHeader.Priority),
-		CreateTime:      createTime,
-		BypassAdmission: bypassAdmission,
-	}
-
-	admissionEnabled := true
-	// Don't subject HeartbeatTxnRequest to the storeAdmissionQ. Even though
-	// it would bypass admission, it would consume a slot. When writes are
-	// throttled, we start generating more txn heartbeats, which then consume
-	// all the slots, causing no useful work to happen. We do want useful work
-	// to continue even when throttling since there are often significant
-	// number of tokens available.
-	if ba.IsWrite() && !ba.IsSingleHeartbeatTxnRequest() {
-		storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(ba.Replica.StoreID))
-		if storeAdmissionQ != nil {
-			storeWorkHandle, err := storeAdmissionQ.Admit(
-				ctx, admission.StoreWriteWorkInfo{WorkInfo: admissionInfo})
-			if err != nil {
-				return AdmissionHandle{}, err
-			}
-			admissionEnabled = storeWorkHandle.AdmissionEnabled()
-			if admissionEnabled {
-				defer func() {
-					if retErr != nil {
-						// No bytes were written.
-						_ = storeAdmissionQ.AdmittedWorkDone(ah.storeWorkHandle, admission.StoreWorkDoneInfo{})
-					}
-				}()
-				ah.storeAdmissionQ, ah.storeWorkHandle = storeAdmissionQ, storeWorkHandle
-			}
-		}
-	}
-	if admissionEnabled {
-		if ba.IsSingleExportRequest() {
-			// Backups generate batches with single export requests, which we
-			// admit through the elastic CPU work queue. We grant this
-			// CPU-intensive work a set amount of CPU time and expect it to
-			// terminate (cooperatively) once it exceeds its grant. The amount
-			// disbursed is 100ms, which we've experimentally found to be long
-			// enough to do enough useful work per-request while not causing too
-			// much in the way of scheduling delays on individual cores. Within
-			// admission control we have machinery that observes scheduling
-			// latencies periodically and reduces the total amount of CPU time
-			// handed out through this mechanism, as a way to provide latency
-			// isolation to non-elastic ("latency sensitive") work running on
-			// the same machine.
-			elasticWorkHandle, err := n.elasticCPUWorkQueue.Admit(
-				ctx, elasticCPUDurationPerExportRequest.Get(&n.settings.SV), admissionInfo,
-			)
-			if err != nil {
-				return AdmissionHandle{}, err
-			}
-			ah.ElasticCPUWorkHandle = elasticWorkHandle
-			defer func() {
-				if retErr != nil {
-					// No elastic work was done.
-					n.elasticCPUWorkQueue.AdmittedWorkDone(ah.ElasticCPUWorkHandle)
-				}
-			}()
-		} else {
-			callAdmittedWorkDoneOnKVAdmissionQ, err := n.kvAdmissionQ.Admit(ctx, admissionInfo)
-			if err != nil {
-				return AdmissionHandle{}, err
-			}
-			ah.callAdmittedWorkDoneOnKVAdmissionQ = callAdmittedWorkDoneOnKVAdmissionQ
-		}
-	}
-	return ah, nil
-}
-
-// AdmittedKVWorkDone implements the KVAdmissionController interface.
-func (n *KVAdmissionControllerImpl) AdmittedKVWorkDone(
-	ah AdmissionHandle, writeBytes *StoreWriteBytes,
-) {
-	n.elasticCPUWorkQueue.AdmittedWorkDone(ah.ElasticCPUWorkHandle)
-	if ah.callAdmittedWorkDoneOnKVAdmissionQ {
-		n.kvAdmissionQ.AdmittedWorkDone(ah.tenantID)
-	}
-	if ah.storeAdmissionQ != nil {
-		var doneInfo admission.StoreWorkDoneInfo
-		if writeBytes != nil {
-			doneInfo = admission.StoreWorkDoneInfo(*writeBytes)
-		}
-		err := ah.storeAdmissionQ.AdmittedWorkDone(ah.storeWorkHandle, doneInfo)
-		if err != nil {
-			// This shouldn't be happening.
-			if buildutil.CrdbTestBuild {
-				log.Fatalf(context.Background(), "%s", errors.WithAssertionFailure(err))
-			}
-			if n.every.ShouldLog() {
-				log.Errorf(context.Background(), "%s", err)
-			}
-		}
-	}
-}
-
-// SetTenantWeightProvider implements the KVAdmissionController interface.
-func (n *KVAdmissionControllerImpl) SetTenantWeightProvider(
-	provider TenantWeightProvider, stopper *stop.Stopper,
-) {
-	// TODO(irfansharif): Use a stopper here instead.
-	go func() {
-		const weightCalculationPeriod = 10 * time.Minute
-		ticker := time.NewTicker(weightCalculationPeriod)
-		// Used for short-circuiting the weights calculation if all weights are
-		// disabled.
-		allWeightsDisabled := false
-		for {
-			select {
-			case <-ticker.C:
-				kvDisabled := !admission.KVTenantWeightsEnabled.Get(&n.settings.SV)
-				kvStoresDisabled := !admission.KVStoresTenantWeightsEnabled.Get(&n.settings.SV)
-				if allWeightsDisabled && kvDisabled && kvStoresDisabled {
-					// Have already transitioned to disabled, so noop.
-					continue
-				}
-				weights := provider.GetTenantWeights()
-				if kvDisabled {
-					weights.Node = nil
-				}
-				n.kvAdmissionQ.SetTenantWeights(weights.Node)
-				n.elasticCPUWorkQueue.SetTenantWeights(weights.Node)
-
-				for _, storeWeights := range weights.Stores {
-					q := n.storeGrantCoords.TryGetQueueForStore(int32(storeWeights.StoreID))
-					if q != nil {
-						if kvStoresDisabled {
-							storeWeights.Weights = nil
-						}
-						q.SetTenantWeights(storeWeights.Weights)
-					}
-				}
-				allWeightsDisabled = kvDisabled && kvStoresDisabled
-			case <-stopper.ShouldQuiesce():
-				ticker.Stop()
-				return
-			}
-		}
-	}()
-}
-
-// SnapshotIngested implements the KVAdmissionController interface.
-func (n *KVAdmissionControllerImpl) SnapshotIngested(
-	storeID roachpb.StoreID, ingestStats pebble.IngestOperationStats,
-) {
-	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(storeID))
-	if storeAdmissionQ == nil {
-		return
-	}
-	storeAdmissionQ.StatsToIgnore(ingestStats)
-}
-
-// FollowerStoreWriteBytes implements the KVAdmissionController interface.
-func (n *KVAdmissionControllerImpl) FollowerStoreWriteBytes(
-	storeID roachpb.StoreID, followerWriteBytes followerStoreWriteBytes,
-) {
-	if followerWriteBytes.WriteBytes == 0 && followerWriteBytes.IngestedBytes == 0 {
-		return
-	}
-	storeAdmissionQ := n.storeGrantCoords.TryGetQueueForStore(int32(storeID))
-	if storeAdmissionQ == nil {
-		return
-	}
-	storeAdmissionQ.BypassedWorkDone(
-		followerWriteBytes.numEntries, followerWriteBytes.StoreWorkDoneInfo)
-}
-
-// ProvisionedBandwidthForAdmissionControl set a value of the provisioned
-// bandwidth for each store in the cluster.
-var ProvisionedBandwidthForAdmissionControl = settings.RegisterByteSizeSetting(
-	settings.SystemOnly, "kv.store.admission.provisioned_bandwidth",
-	"if set to a non-zero value, this is used as the provisioned bandwidth (in bytes/s), "+
-		"for each store. It can be over-ridden on a per-store basis using the --store flag",
-	0).WithPublic()
